C
A 2km road with a 7.5 m width and four vehicle classes are simulated using python for testing model properties. The simulation was done for 30 minutes. The road was divided into several grids, with 100 grids along the length of the road and 10 grids along the width of the road. Density fields, class-specific flux density, and velocity fields are initialized with random values scaled to average densities and velocities, respectively. The time step was calculated to satisfy the CFL conditions and ensure stability.